python/recall/requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
MarkupSafe==1.1.1
mongoengine==0.20.0
opentelemetry-api==0.12b0
opentelemetry-exporter-jaeger==0.12b0
opentelemetry-instrumentation==0.12b0
opentelemetry-instrumentation-flask==0.12b0
opentelemetry-instrumentation-pymongo==0.12b0
opentelemetry-instrumentation-wsgi==0.12b0
opentelemetry-sdk==0.12b0
Pillow==7.2.0
prometheus-client==0.8.0
pymongo==3.11.0
python-magic==0.4.18
python-magic-bin==0.4.14
six==1.15.0
thrift==0.13.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
zipp==3.1.0
